// components/api/openapi/models/PaginatedRecordsDTOSirutaDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SirutaDTO } from './SirutaDTO';
export type PaginatedRecordsDTOSirutaDTO = {
    page?: number;
    total?: number;
    records?: number;
    rows?: Array<SirutaDTO>;
};